FromJDA refactor
