+++ src/main/java/net/dv8tion/jda/api/entities/Mentions.java
@@ -17,7 +17,7 @@
 package net.dv8tion.jda.api.entities;
 
 import net.dv8tion.jda.api.entities.emoji.CustomEmoji;
-import net.dv8tion.jda.api.from.FromJDA;
+import net.dv8tion.jda.api.FromJDA;
 import org.apache.commons.collections4.Bag;
 
 import javax.annotation.Nonnull;
