+++ src/main/java/net/dv8tion/jda/api/events/GenericEvent.java
@@ -16,7 +16,7 @@
 
 package net.dv8tion.jda.api.events;
 
-import net.dv8tion.jda.api.from.FromJDA;
+import net.dv8tion.jda.api.FromJDA;
 import net.dv8tion.jda.api.utils.data.DataObject;
 
 import javax.annotation.Nullable;
