--- src/main/java/net/dv8tion/jda/api/from/FromJDA.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 2015 Austin Keener, Michael Ritter, Florian Spieß, and the JDA contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package net.dv8tion.jda.api.from;
-
-import net.dv8tion.jda.api.JDA;
-
-import javax.annotation.Nonnull;
-
-/**
- * An interface for objects that have a corresponding jda instance
- */
-public interface FromJDA
-{
-    /**
-     * The corresponding {@link net.dv8tion.jda.api.JDA jda} instance
-     *
-     * @return The jda instance
-     */
-    @Nonnull
-    JDA getJDA();
-
-}
