+++ src/main/java/net/dv8tion/jda/api/utils/SessionController.java
@@ -17,7 +17,7 @@
 package net.dv8tion.jda.api.utils;
 
 import net.dv8tion.jda.api.JDA;
-import net.dv8tion.jda.api.from.FromJDA;
+import net.dv8tion.jda.api.FromJDA;
 
 import javax.annotation.Nonnull;
